+++ memtable-core/src/impls/mem/fixed/mod.rs
@@ -2,7 +2,6 @@ mod both;
 mod col;
 mod row;
 
-use super::utils;
 pub use both::FixedTable;
 pub use col::FixedColumnTable;
 pub use row::FixedRowTable;