+++ memtable-macros/tests/derive/table.rs
@@ -376,7 +376,7 @@ fn should_support_trying_to_convert_from_untyped_table() {
     // If data is in right columns and is not missing anything,
     // conversion should work fine
     {
-        let mut table = memtable::MemTable::new();
+        let mut table = memtable::MemDynamicTable::new();
         table.push_row(vec![
             MyRowTableData::Field1(false),
             MyRowTableData::Field2(123),
@@ -395,7 +395,7 @@ fn should_support_trying_to_convert_from_untyped_table() {
 
     // If data is not in right order in terms of column types, should fail
     {
-        let mut table = memtable::MemTable::new();
+        let mut table = memtable::MemDynamicTable::new();
         table.push_row(vec![
             MyRowTableData::Field2(123),
             MyRowTableData::Field1(false),
@@ -409,7 +409,7 @@ fn should_support_trying_to_convert_from_untyped_table() {
 
     // If data is missing in places, should fail
     {
-        let mut table = memtable::MemTable::new();
+        let mut table = memtable::MemDynamicTable::new();
         table.push_row(vec![
             MyRowTableData::Field1(false),
             MyRowTableData::Field2(123),